refactor(telegram): move configuration types into telegram package
