+++ config/notifiers.go
@@ -143,15 +143,6 @@ var (
 		Subject: `{{ template "sns.default.subject" . }}`,
 		Message: `{{ template "sns.default.message" . }}`,
 	}
-
-	DefaultTelegramConfig = TelegramConfig{
-		NotifierConfig: amcommoncfg.NotifierConfig{
-			VSendResolved: true,
-		},
-		DisableNotifications: false,
-		Message:              `{{ template "telegram.default.message" . }}`,
-		ParseMode:            "HTML",
-	}
 )
 
 // WebexConfig configures notifications via Webex.
@@ -654,48 +645,6 @@ func (c *SNSConfig) UnmarshalYAML(unmarshal func(any) error) error {
 	}
 	if (c.TargetARN == "") != (c.TopicARN == "") != (c.PhoneNumber == "") {
 		return errors.New("must provide either a Target ARN, Topic ARN, or Phone Number for SNS config")
-	}
-	return nil
-}
-
-// TelegramConfig configures notifications via Telegram.
-type TelegramConfig struct {
-	amcommoncfg.NotifierConfig `yaml:",inline" json:",inline"`
-
-	HTTPConfig *commoncfg.HTTPClientConfig `yaml:"http_config,omitempty" json:"http_config,omitempty"`
-
-	APIUrl               *amcommoncfg.URL `yaml:"api_url" json:"api_url,omitempty"`
-	BotToken             commoncfg.Secret `yaml:"bot_token,omitempty" json:"token,omitempty"`
-	BotTokenFile         string           `yaml:"bot_token_file,omitempty" json:"token_file,omitempty"`
-	ChatID               int64            `yaml:"chat_id,omitempty" json:"chat,omitempty"`
-	ChatIDFile           string           `yaml:"chat_id_file,omitempty" json:"chat_file,omitempty"`
-	MessageThreadID      int              `yaml:"message_thread_id,omitempty" json:"message_thread_id,omitempty"`
-	Message              string           `yaml:"message,omitempty" json:"message,omitempty"`
-	DisableNotifications bool             `yaml:"disable_notifications,omitempty" json:"disable_notifications,omitempty"`
-	ParseMode            string           `yaml:"parse_mode,omitempty" json:"parse_mode,omitempty"`
-}
-
-// UnmarshalYAML implements the yaml.Unmarshaler interface.
-func (c *TelegramConfig) UnmarshalYAML(unmarshal func(any) error) error {
-	*c = DefaultTelegramConfig
-	type plain TelegramConfig
-	if err := unmarshal((*plain)(c)); err != nil {
-		return err
-	}
-	if c.BotToken != "" && c.BotTokenFile != "" {
-		return errors.New("at most one of bot_token & bot_token_file must be configured")
-	}
-	if c.ChatID == 0 && c.ChatIDFile == "" {
-		return errors.New("missing chat_id or chat_id_file on telegram_config")
-	}
-	if c.ChatID != 0 && c.ChatIDFile != "" {
-		return errors.New("at most one of chat_id & chat_id_file must be configured")
-	}
-	if c.ParseMode != "" &&
-		c.ParseMode != "Markdown" &&
-		c.ParseMode != "MarkdownV2" &&
-		c.ParseMode != "HTML" {
-		return errors.New("unknown parse_mode on telegram_config, must be Markdown, MarkdownV2, HTML or empty string")
 	}
 	return nil
 }
